chore: add searchpage to integration
