+++ web/src/locales/langs/en-us/page.ts
@@ -171,6 +171,8 @@ const page: App.I18n.Schema['translation']['page'] = {
         feature_think_active: 'Enable Deep Think',
         hotkey: 'Hotkey',
         hotkey_placeholder: 'Set hotkey to call out CocoAI',
+        module_ai_widgets: 'AI Widgets',
+        module_ai_overview: 'AI Overview',
         module_chat: 'AI Chat',
         module_chat_ai_assistant: 'AI Assistant',
         module_chat_placeholder: 'Chat box placeholder text',
@@ -183,13 +185,16 @@ const page: App.I18n.Schema['translation']['page'] = {
         theme_dark: 'Dark',
         theme_light: 'Light',
         type: 'Type',
-        type_all: 'All',
-        type_embedded: 'Embedded',
-        type_embedded_placeholder: 'Embedded widget placeholder text',
-        type_embedded_icon: 'Embedded widget icon',
-        type_floating: 'Floating',
-        type_floating_placeholder: 'Floating widget placeholder text',
-        type_floating_icon: 'Floating widget icon',
+        type_searchbox: 'Search Box',
+        type_searchpage: 'Search Page',
+        mode: 'Mode',
+        mode_all: 'All',
+        mode_embedded: 'Embedded',
+        mode_embedded_placeholder: 'Embedded widget placeholder text',
+        mode_embedded_icon: 'Embedded widget icon',
+        mode_floating: 'Floating',
+        mode_floating_placeholder: 'Floating widget placeholder text',
+        mode_floating_icon: 'Floating widget icon',
       },
       title: {
         edit: 'Edit Integration',
