+++ web/src/locales/langs/zh-cn/page.ts
@@ -166,6 +166,8 @@ const page: App.I18n.Schema['translation']['page'] = {
         feature_think_active: '开启深度思考',
         hotkey: '快捷键',
         hotkey_placeholder: '设置呼出 CocoAI 的快捷键',
+        module_ai_widgets: 'AI 组件',
+        module_ai_overview: 'AI Overview',
         module_chat: 'AI 聊天',
         module_chat_ai_assistant: 'AI 助手',
         module_chat_placeholder: 'AI 聊天输入框提示文本',
@@ -178,13 +180,15 @@ const page: App.I18n.Schema['translation']['page'] = {
         theme_dark: '深色',
         theme_light: '浅色',
         type: '类型',
-        type_all: '内嵌和浮动',
-        type_embedded: '内嵌的',
-        type_embedded_placeholder: '内嵌组件提示文本',
-        type_embedded_icon: '内嵌组件图标',
-        type_floating: '浮动的',
-        type_floating_placeholder: '浮动组件提示文本',
-        type_floating_icon: '浮动组件图标',
+        type_searchbox: '搜索框',
+        type_searchpage: '搜索页',
+        mode_all: '内嵌和浮动',
+        mode_embedded: '内嵌的',
+        mode_embedded_placeholder: '内嵌组件提示文本',
+        mode_embedded_icon: '内嵌组件图标',
+        mode_floating: '浮动的',
+        mode_floating_placeholder: '浮动组件提示文本',
+        mode_floating_icon: '浮动组件图标',
       },
       title: {
         edit: '编辑嵌入组件',
